runtime: Updated SIGSYS to cause a panic + stacktrace.
On GNU/Linux, SIGSYS is specified to cause the process to terminate without
a core dump. In https://codereview.appspot.com/3749041 , it appears that Golang
accidentally introduced incorrect behavior for this signal, which caused Goalng
processes to keep running after receiving SIGSYS. This change reverts it to the
old/correct behavior.

Fixes #15204

Change-Id: I3aa48a9499c1bc36fa5d3f40c088fdd7599e0db5