Update patch set 1
Patch Set 1: Code-Review+2

Yay deleting!

For the record,

https://msdn.microsoft.com/en-us/library/windows/desktop/ms738520(v=vs.85).aspx

says:

"The GetAddrInfoW function is the Unicode version of getaddrinfo. The GetAddrInfoW function was added to the Ws2_32.dll in Windows XP with Service Pack 2 (SP2). The GetAddrInfoW function cannot be used on versions of Windows earlier than Windows XP with SP2."

So I guess our minimum requirements is XP Service Pack 2.

I'm okay with that.

Patch-set: 1
Label: Code-Review=+2