Update patch set 4
Patch Set 4:

> It's in scanstack. We could certainly lift that check higher, which would shave off time. But a non-trivial amount of that cost is taking the cache miss just to get the flag field out of the g and that cost will be there with any flag-based approach.

You are right about the cache-miss.

> Where does the 100ns come from? That sounds like a microbenchmark that does nothing but start and exit goroutines.

That's close to a synthetic benchmark (synthetic benchmark can do it even faster).    Russ says that slowing down synthetic benchmarks like sieve and threadring is no-no. Also for real apps it's just a matter of running on enough cores. E.g. if goroutine runs for 100us and we have 64 cores, the mutex will be a serious issue.

Patch-set: 4