Update patch set 2
Patch Set 2:

> Are there 3 implementations possible here? Pure Go,
 > s390x assembly, and s390x+VX instructions?

There are only 2 implementations, pure Go and s390x+VX. The pure Go implementation is used when the vector instructions are not available or the data length is less than 64 bytes. The pure Go implementation is also used to align the data if necessary.

Chris, do you want to add this info to the commit message? The s390x side of this LGTM.

Patch-set: 2
Label: Code-Review=0