Update patch set 2
Patch Set 2:

> > (12 comments)
 > >
 > > > (6 comments)
 > > >
 > > > Can you make sure the tests are testing both the VX and non-VX
 > > code
 > > > paths?
 > >
 > > Yes, the non-VX code path is taken when the data length is below
 > 64
 > > bytes.
 > 
 > I mean that for every test input, we want to test every possible
 > implementation. Are there 3 implementations possible here? Pure Go,
 > s390x assembly, and s390x+VX instructions?

Hi Brad,
The implementation processes parts of the input data buffer with the pure go routine. The input buffer is aligned to a doubleword boundary to improve performance of the Vector instruction.

 > > (12 comments)
 > >
 > > > (6 comments)
 > > >
 > > > Can you make sure the tests are testing both the VX and non-VX
 > > code
 > > > paths?
 > >
 > > Yes, the non-VX code path is taken when the data length is below
 > 64
 > > bytes.
 > 
 > I mean that for every test input, we want to test every possible
 > implementation. Are there 3 implementations possible here? Pure Go,
 > s390x assembly, and s390x+VX instructions?

Hi Brad,
As Michael said, Pure Go is used in vectorizedCRC32 algorithm to align the input buffer because vector instructions are used to process a multiple of 128-bit data chunks to avoid a padding. Also, shorter buffer less than 64 bytes are processed with pure go.

Patch-set: 2
Label: Code-Review=0