Update patch set 1
Patch Set 1:

Hmm. The expensive thing here is the division, and using the maxElem lookup from the other CL could get rid of one of two divisions. So I think I'll abandon this and add that to the other CL. That seems like a clear, easy win.

I'm not convinced yet that we shouldn't still add other cases here, but I'm not sure I want to spend the time now fiddling with benchmarks.

Patch-set: 1