Update patch set 4
Patch Set 4:

Actually, I don't really need to remove the G from the rescan list when it exits. Then I also don't have to re-add it if it gets recycled. That slightly increases the size of the rescan list (since it will include Gs that ran and exited and haven't been reused), but should significantly reduce contention on the lock.

Patch-set: 4