Update patch set 2
Patch Set 2:

> Thanks for doing this. I tried to be too clever when I tried this
 > and got bogged down in walk.go.
 > 
 > Given how often it occurs, would it be worth adding
 > 
 > var zerop = unsafe.Pointer(&zeroVal[0])
 > 
 > and returning that everywhere?
 > 

That would require a load instead of a LEAQ, so I don't think it is worth switching.

 > Maybe a test in /test is called for? IIRC this code path is not
 > actually hit anywhere in ./all.bash, but maybe I'm wrong (if you
 > reduce mapZero to under 1000 then test/bigmap.go will test this
 > code path...)
 > 

Added a test.

 > Also I wonder if maxZero could be made smaller, did you think much
 > about the value to use?

I didn't really give it much thought.  1K seemed reasonable, and is the current setting for "really big".  I'm happy to use another number if you think it would be better.

Patch-set: 2