runtime,reflect: use uintptr type consistently for newarray
Synchronize the type of the length argument of
runtime.reflect_unsafe_NewArray and reflect.unsafe_NewArray
to uintptr.

Remove the int(n) < 0 check from newarray since the type of
n is uintptr and all callers are responsible and do make sure
that no negative int is cast and passed as uintptr into newarray.

Change-Id: Ic77bf2dde216c92ca8c49462f8eedc0385b6314e