Update patch set 8
Patch Set 8:

> @Nigel - When I look at the modified Snappy code, I cannot see why
 > re-implementing it should yield any improvements, and considering
 > the amount of changes I have done to adapt it to deflate I would
 > rather stick with what we have, at least for now.

I re-based on the "new Snappy algorithm", aka the same algorithm that the C++ Snappy code uses.

Call what this CL implements the "old Snappy algorithm".

The bottom line is that, on my measurements, "new Snappy" has both higher throughput and smaller output than "old Snappy". The details are at https://go-review.googlesource.com/#/c/22369/

That CL isn't ready to submit as is. There are tests to write and comments to add, but I published it in order to share the numbers and the code.

Moving forward, my preference would be to land "new Snappy", but as I said before, to do so in stages. The first stage would be `replace "Best Speed" with a not-as-clever-but-more-obviously-correct Snappy style algorithm` which is more or less https://go-review.googlesource.com/#/c/22369/

How would you like to proceed? I am happy for you to have Author attribution if we go with "new Snappy".

Patch-set: 8