Update patch set 1
Patch Set 1: Code-Review+2

(I'd call the function Load.)> Patch Set 1:
> 
> > Your new []COFFSymbol is much more memory efficient than the old
>  > []*Symbol. It seems a pity to box yourself into generating the old
>  > []*Symbol on read.
> 
> I was thinking the same. Especially since we're going to use this in cmd/link, where every little bit counts.
> 
> My current godoc executable file has 15459 pe symbols. Is that a lot?

Go programs 10x larger than godoc are not uncommon and worth designing for in the linker. Assuming linear scaling of symbols (which seems to be true in the programs I've looked at), thinking on the order of 200k symbols makes sense to me.

> And average Windows executable will have no symbols in their symbol table. We append symbols to every Go executable to make objdump and addr2line work. But other executables don't have that requirement.
> 
> All PE object files have symbols. And that is where cmd/link is affected - reading PE object files generated by gcc (for cgo only). How much slower this process becomes because of extra []*Symbol reading? Perhaps we should wait and see.
> 
>  > 
>  > What do you think of introducing a new load function, that takes as
>  > an option whether it generates the []*Symbol? The old Open/NewFile
>  > can call the new load function with it set to true.
> 
> Sounds like a good suggestion. But once you start thinking of a good name for that. Do you have a good name for it? And trying to explain why you need this extra function (perhaps even 2 for both Open and NewFile). And the reason is performance only. And do you propose to leave File.Symbol empty if new function parameter says so? It is all feels bit too fiddly.

I was thinking Load, as a replacement for NewFile. The Open function is unfortunate to me, I don't see a tiny helper like that being worth dragging in the os package, so I don't think it's worth worrying about. Let people write that themselves.

> I think regardless what we decide here, we should proceed with this CL as is. Whatever things this CL exports will need to be exported regardless. Don't you think? Anything else we can adjust later.
> 
> I also feel we'll run out of time before 1/5/2016 freeze. I will probably unexport all new exported things for the freeze period, and revert this once go1.8 development starts. I don't want to commit to new API until we've tested it with cmd/link.

I'm OK with this CL if you're going to unexport all of this for 1.7. I agree that it's a little late in the cycle to be designing this API.

> Alex

Patch-set: 1
Reviewer: Gerrit User 5030 <5030@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2