Update patch set 10
Patch Set 10:

(3 comments)

There are still a bunch of (s.start << _PageShift)s. Why not just replace them all while we're here?

(For future reference, the cleaner way to structure a change like this is as two CLs: one that replaces all uses of (s.start << _PageShift) with s.base and another that does the change to s.base.)

Patch-set: 10
Label: Code-Review=0