Update patch set 1
Patch Set 1:

As I see it, this CL allows you to discover why something is stale before you realize it's style because of all the rebuilding. In short, how is one expected to use this feature?

Wouldn't it make more sense if an existing flag, such as -v or -x, included this information by default?

Patch-set: 1