Update patch set 1
Patch Set 1:

> As I see it, this CL allows you to discover why something is stale
 > before you realize it's style because of all the rebuilding. In
 > short, how is one expected to use this feature?
 > 
 > Wouldn't it make more sense if an existing flag, such as -v or -x,
 > included this information by default?

I think this CL should go in regardless (more info exposed in go list helps tools that are built on top of cmd/go and also helps people debugging cmd/go). But I agree that it might also make sense to show the info in the -v or -x prints. I'll look into that in a followup.

Patch-set: 1