cmd/compile: fix another bug in dominator computation
Here, "fix" means "replace".  The new dominator computation
is the "simple" algorithm from Lengauer and Tarjan's TOPLAS
paper, with minimal changes.

Also included is a test that tweaks the fixed error; it was
from a piece of code in the existing runtime (!) and a new
function to automate creation of canned tests.

Change-Id: I0abdf53d5d64df1e67e4e62f55e88957045cd63b