Update patch set 3
Patch Set 3:

I'm still hitting this same error on the latest master when I pick a goroutine from the goroutine analysis list. I'm trying to look at some fairly large traces, so it may also be triggered by a timeout in the XHR call, although the error seems to coincide pretty closely with the server side logging successful transmission of the trace to the browser (I added the logging output to the trace binary).

$ ../pprof/trace rdb rdb-profile833129407
2016/04/25 16:26:43 parsing rdb-profile833129407
2016/04/25 16:29:50 parsed 44363245 events
2016/04/25 16:30:30 tracing 44363245 events
2016/04/25 16:30:51 traced 43138 events, 608 frames
2016/04/25 16:30:52 sent serialized trace

Patch-set: 3