Update patch set 5
Patch Set 5:

Converting a string into a byte slice as done here is unfortunately a novel pattern in the codebase, as far as I can tell. Switching to ptr+len feels slightly more honest to me (and could be a bit more performant), but would require some more code duplication and would make the generic versions in Go more awkward. I'm happy to pursue that, but it isn't the least invasive route.

Patch-set: 5