Update patch set 1
Patch Set 1:

> Patch Set 1: -Run-TryBot
> 
> > I see that SIGSYS is SigUnblock on GNU/Linux, but actually I don't see why that is correct. 
> 
> wait, you gave a sign-off to that change, right?

Yes, but I probably made a mistake.  At the time I didn't notice that it was adding SigUnblock.  I should have looked more closely.

> > What is the reason for this change?  What does it fix?
> 
> my interpretation of https://go-review.googlesource.com/22202 is like the following:
> - the expected behavior (and probably bottom consensus) on sigsys and friends btw bsd/svr/posix is simply "terminate the process as soon as possible",
> - on linux, sigsys has a fancy feature that kernel doesn't generate a core image, probably other platforms don't have it, but it's okay,
> - we don't unblock sigsys for the main purpose like its friends such as sigbus, sigill, etc.
> 
> this change makes sigsys effective at anywhere not only on linux but other sigunblock supported platforms.

The meaning of the SigUnblock flag is that the Go runtime will ensure that the signal is not blocked.  We set it for signals that the program must catch in order to work correctly.  It's not obvious to me that SIGSYS is such a signal.  SIGSYS is not really like SIGBUS.

Patch-set: 1
Label: TryBot-Result=0