Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> How does this change compare to Russ's comment in https://github.com/golang/go/issues/14232#issuecomment-180119370 ?

It's cheap enough that it doesn't hurt to provide the extra functionality.
Late in the 1.6 cycle we needed an obviously safe fix.  That doesn't apply now.

I also noticed that if the []byte escapes, we do (and always did?) return a []byte which is larger than the string.  So I don't think we were ever always returning cap==len(string) for string->[]byte conversions.

Patch-set: 2