Update patch set 5
Patch Set 5:

What was the reason for not putting the Context in the Cmd struct here? The Cmd seems very clearly a 'parameter struct' (see my comment in #14660). I am worried that now we have RunContext and WaitContext but we're missing OutputContext and CombinedOutputContext and so on. Because the Cmd is a parameter struct and Cmds are not reusable already, putting the Context into the Cmd would make it apply to all these operations with no new methods.

Patch-set: 5