Update patch set 10
Patch Set 10:

> Scott, it might help to look at how this is done eg for bytes.Equal
 > and runtime.memeqbody in runtime/asm_amd64.s.

Thanks, Josh. Using memeqbody as an example, it seems to me that what we're doing here for amd64 is the same: the entry point declares the parameter size, moves those into known registers and JMPs to the common routine, which is declared with no parameters and no stack space (all NOSPLIT.) We're never CALLing the shared routine, nor does it make reference to FP, so from what I can see we're playing nicely. I will update block<> to spell out the parameter size of zero, since that seems to be done elsewhere ($0-0.)

On 386, I currently have a dubious reference to FP in block<> to refer back to the digest once the hashing is finished. If that would be problematic for unwinding or some other case, the thing to do might be to stash that in an additional stack slot. 

On arm, there is the same dubious FP reference in block<>. Interestingly, bytes.Equal, memequal, and eqstring don't share a common routine in asm_arm.s, though there is a TODO comment suggesting it. cmpstring and bytes.Compare do jump to runtime·cmpbody(SB) in the same fashion, though, and those entry points as well as cmpbody declare $-4 so as not to save LR, which I think we should be doing as well.

On ppc64, there's the FP reference in block<> and the wrong mnemonic for jumping, and it looks like I want NOFRAME as well. 

An s390 implementation popped up in the meanwhile, so I'll include support for that. 

Lastly, it looks like I need to switch to CALL in amd64p32 to get automatic stack alignment.

Does all of that sound accurate? I can make these changes, run the profiler on amd64 and 386, and update the md5 benchmarks (Write and WriteString) before we decide to proceed further.

Patch-set: 10