Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> My concern may be unfounded.
> 
> @khr: can you explain how we unwind the stack if we receive a profiling signal while executing within memeqbody in runtime/asm_amd64.s?

Unwinding should work fine.  Yes, it is going to think that there are no arguments.  For stack unwinding for profiling, that is ok.
It would not be ok if memeqbody had a GC safepoint in it.  Then the existence and layout of the args section matters.

I guess we also need the "callers" of memeqbody (memequal, memequal_varlen, eqstring, and bytes.Equal) to have no stack frame.  Then only the args are different, the frame itself is identical.

Patch-set: 10