runtime: simplify heapBitsSetType
DO NOT REVIEW. I expected this to speed things up, but it doesn't. I'm
posting it to Gerrit for posterity.

Now that the scan/dead bit in the first heap word of an object isn't
special, there's no need to special case heapBitsSetType when the
first word is bitmap-aligned. Hence, remove this special case path.

However, this actually slows things down a bit. My theory is that,
because of this special case, we handle aligned objects up to 8 words
with a shorter path through the code, but without this special case,
this shorter path only applies up to 4 words.

Change-Id: I87571d7e499692bc0dcedb31e28db3b6e1bf5909