Update patch set 14
Patch Set 14:

> ppc64 has the same problem :) I don't mind fixing this if you'd like, I have access to a ppc64 machine so I can test it.

If you could test it, that would be great.  Thanks.  I thought I understand the issue on arm64 but my understanding does not translate to PPC.

Note that I think that I botched the argument size on PPC for cgocallback_gofunc, but I don't think that is the core problem.

I see that runtime/cgo/asm_ppc64.s is obviously horribly wrong, I had not noticed that before.

Patch-set: 14