Update patch set 1
Patch Set 1:

> (1 comment)

correct. multiple libgcc_s was the problem on the FreeBSD 10.2 machine.

The problem is that darwin fails with rpath set:
ld: -rpath can only be used when creating a dynamic final linked image

so i guess the more immediate universal solution is to leave rpath out for now.

Patch-set: 1