Update patch set 2
Patch Set 2:

> You meaned: is a user responsible for solving correct library
 > order?
 > I don't know whitch is better or not.

Seems a problem with darwin and rpath being handled special there. So i leave rpath out on osx for now since im not sure how it should be handled there and it seems better to fix the library path issue at any rate independently of a possible rpath issue on darwin. At least this way it works on darwin when only one gcc is installed for now.

Patch-set: 2