Update patch set 2
Patch Set 2:

This improves the situation, but in my opinion it does not resolve #15399. The fallacy described in that issue is still quite likely: I can have a "trusted source" which always gives me valid JSON and not some bad JS. I could even see that it gets "included verbatim" in the output. Still, it is not trivially apparent to me that the >content of a string< inside of the JSON can break my neck.

Patch-set: 2
Reviewer: Gerrit User 13620 <13620@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0