cmd/compile: expand sparseset for regalloc to prevent index error
In regalloc, a sparse map is preallocated for later use by
spill-in-loop sinking.  However, variables (spills) are added
during register allocation before spill sinking, and a map
query involving any of these new variables will index out of
bounds in the map.

To fix:
1) fix the queries to use s.orig[v.ID].ID instead, to ensure
proper indexing.  Note that s.orig will be nil for values
that are not eligible for spilling (like memory).

2) modify the sparse set&map code to use a slice of int32
for their index data structure, because in practice if there's
more than 2 billion blocks or variables, other problems will
bite first.

3) slightly harden the sparse set&map code -- out-of-bounds
removes are no-ops, since clearly there's no entry for the
value (better that it be brittle, or fail faster?)

4) add a test.

Fixes #15585.

Change-Id: I8f2caa93b132a0f2a9161d2178320d5550583075