Update patch set 1
Patch Set 1: Code-Review-1

I don't like this fix.  It's fixing a symptom, not the root problem.  The root problem is you're looking things up using the a spill or restore value instead of the original (pre-regalloc) value.  Just making the array larger doesn't change the fact that you're looking up the wrong index.  It just means you're getting the wrong value instead of a bounds error.  The s.orig array is for exactly this problem, mapping from a spill/restore of a value back to the original value.  Judicious use of s.orig should fix this problem correctly.

The change in sparsemap/set to use int32 instead of int is fine.  Maybe it should wait for 1.8 though?

Patch-set: 1
Label: Code-Review=-1