Update patch set 2
Patch Set 2:

s.orig indexing is now in, doubling is gone.

I'd love to keep the slightly smaller sets in, given how memory pressure is still an issue for some compilations, and it's one of the safest changes I know of.  It's worth a percent or so in the gogo/protobuf stuff even with sparse phis enabled.

Patch-set: 2