go/importer: use correct path when checking if package was already imported
The importer uses a global (shared) package map across multiple imports
to determine if a package was imported before. That package map is usually
indexed by package path. However, for the package to be imported, the
gcimporter used the file path ('id'). When the file and package path are
the same, this is not noticed. When the file and package path differ, e.g.,
after cleaning up a path containing '..', or a path leading to a vendored
package, the package key is not correct. The textual importer "corrected"
for that mistake by consistently using the file path for the package to be
imported, and when it was fully imported it marked that package as complete.
The binary importer correctly used the actual package path to identify the
package to be imported, and the package marked as complete and entered into
the shared packages map could then not be found using the file path as index.
The next time the same package was imported, instead of bailing out immediately,
the importer attempted to import the same objects again.

To make the story a bit more complicated: The textual exporter also has
this problem if imported objects use a package path that is different from
the file path of the package to be imported. However, the textual importer
doesn't currently check for duplicate imports, which is why this was never
noticed before if the problem occured in the first place. Furthermore,
identical duplicate imports don't lead to errors (the duplicates are simply
discarded); but there is a performance issue.

In this fix, use the package path consistently.

In a subsequent change we may want to clean up naming and replace 'id'
with 'path' in gcimporter.go.

Fixes #15517.

Change-Id: I0c32a708dfccf345e0353fbda20ad882121e437c