Redundant condition check
The value of old is being checked against pdReady at the beginning
of the loop in netpollunblock and if equal the function returns.
It is being checked against pdReady again after CAS, which will
never happen since CAS happens only if (*gpp == old) and old is not
modified after the initial check.

Change-Id: Ibb4ac6816f85ff1ee2d0b4eadd7a471d09090f06