runtime: ring buffer for binary debug logging
DO NOT REVIEW

This adds an internal runtime debug log. It uses per-P time-stamped
ring buffers of binary log records. On panic, these buffers are
interleaved and printed.

The debug log is designed so that it can be disabled and compiled away
entirely by building with -tags nodebuglog.

The debug log is also designed so it would be reasonable to decode the logs
from GDB, though this hasn't been implemented.

Change-Id: I1c4ee8ea6b719a4f25bb4cf9e423cce937a14e12