Update patch set 1
Patch Set 1:

Adam, did you had any further thoughts on this?

I feel strongly towards validating the certificate itself when building the chain followed by  target/usage specific checks with instructions (like a fqdn or key-usage) of the service that is using the certificate.

The reasoning behind this is that the certificate should always validate and if it can't be validated it should not be trusted in the first place. This should protect against false issued certificates from a CA that is not authorized to issue certificates for a particular usage or target.

Patch-set: 1