Update patch set 2
Patch Set 2:

No worries, it makes sense to hold off on this until the namespace fixes go in anyways and I didn't really expect this quick hack to work out of the box or fit in with the broader refactor that the XML package probably needs. Let me know if I can be of assistance in this area as the 1.8 milestone kicks off.

Patch-set: 2
Reviewer: Gerrit User 11106 <11106@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0