Update patch set 11
Patch Set 11:

R=go1.8

> Patch Set 11:
> 
> Obviously returning 0, io.EOF is wrong. 
> But lots of readers return 0, nil immediately. 

I'll just fix that for now.

> What's the rationale for breaking that pattern here? Does this come up in some important use case?
> This is basically a feature being introduced.

A few HTTP bugs are blocked on the ability of the net/http package to wait for readability on a socket without consuming anything from it. I can't read a byte and stitch it together later because the Hijacker interface and the Go 1 compatibility promise implicitly says it'll always get back the raw *net.TCPConn.

So yes, I need this feature. And Windows already does it like this, so it's not much of a stretch to make Unix also do it, especially since the (0, io.EOF) behavior was already wrong and the fix was so simple.

> The concrete connection type docs (*TCPConn etc) will need to document it, and net.Conn's interface docs will need to say something like :
> 
>     Implementations may implement a Read requesting 0 bytes by blocking
>     until more than 0 bytes are available or the connection is closed.
> 
> That's a weird thing to say though.

Yes, kinda. But no weirder than all the other 0 byte behavior.

But we can discuss later. Maybe we could just add a TCPConn.WaitReadable() method instead.

Patch-set: 11