Update patch set 1
Patch Set 1: Code-Review+1

(6 comments)

Looks pretty good, thanks for doing all of this.  Are *,/,>>,<< still coming, or are those runtime calls?

I'm surprised there are no changes to the scheduler and/or flagalloc to make sure we don't clobber flags when the "int32+carry"-typed value is live.

Patch-set: 1
Label: Code-Review=+1