Update patch set 1
Patch Set 1:

This approach is simple but it's not clear that it's helpful.  The warning is kind of wishy-washy.  I would rather see something like "profile does not contain symbols and no binary specified."  But there is also the PPROF_BINARY_PATH environment variable to consider, not that I've ever tried using that myself.

Patch-set: 1