Update patch set 2
Patch Set 2:

FWIW, this was not an editing relic. This dates from before the x... argument syntax. Instead of writing

func Print(args ...interface{}) {
    Fprint(os.Stdout, args...)
}

you wrote

func Print(args ...interface{}) {
    Fprint(os.Stdout, args)
}

The magic was that although args declared as ...interface{}
ostensibly behaved like a []interface{},
in the special case of passing it as the sole entry in another
call's ... argument it got passed through directly.
In modern terms, "args" got turned into "args..." automatically
in this one special case, and the explicit "args..." form didn't exist at all.

The original implementation of the function in this CL was:

func join(elem []string, args ...string) string {
    args = elem
    return Join(args)
}

where the expectation was that the caller only ever passed an elem,
never anything for args.

You couldn't Join(elem), because that wouldn't do the passthrough,
but because args was a ...string, you could assign args=elem and
then use Join(args) instead to get the desired effect.

This kind of subterfuge is exactly why we added the more explicit
syntax Join(elem...).

:-)

Patch-set: 2